Changed referrer-policy
Changed referrer-policy from no-referrer to same-origin. It's clearly
not as secure as not sending any referrer at all, but users can run into
issues with services such as vimeo if they're trying to setup content
domain restriction. For the sake of simplicity for users that will be
running default settings we're moving this default to same-origin.