Complete review of code
This release features a major revamp to the PHP section that moves from
PHP 7.0 to PHP 7.2 and integrates OPcache for better performance.

To remove your previous version  of PHP, use the command `sudo apt
remove php7.0*`.

In addition, we performed a major overhaul of all sections of the code.
Some portions of code were moved around to other files to make more
sense in the flow of the overall config.

There were some relatively minor changes to the various nginx values as
well.