Changes to cache variables and vhost layout
Changed the $no_cache variable to $skip_cache. Using $no_cache can be
problematic if a user tries to chance that variable using a find/replace
command, as it may also replace the nginx directive and/or wordpress
cookie that include the no_cache text as well.

Slightly cleaned up the layout of the vhost files to flow better.

Slight change to gzip settings.